Update lets-build-an-add-on.md
updated url for links to json and schema. They all seem to be at the same level, not one level up. Unfortunately new to this md markup so I can't seem to get it to work via the preview presumably because the static site has to be built somehow. As the links are broken at the moment I don't think my edits make anything worse but they may have fixed it :)